requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework==3.13.1
pandas
django-filter
requests
debugpy
coverage
djangorestframework-simplejwt